live calculation of the statistics.
the statistics where calculated after finishing loading the
file/disconnecting from the server. now they can be calculated as soon
as the a scene is instanciated. this is important for connecting to a
server and calculating the statistics as the events occur.

the udp reader didn't terminate. now it does.